requirements.txt
audio_recorder_streamlit==0.0.8
librosa==0.10.0
numpy==1.22.4
Pillow==10.3.0
streamlit==1.31.1
torch==1.8.1
torchvision==0.9.1
